requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
djangorestframework-recursive>=0.1.2,<0.2.0
django-treenode>=0.16.0,<0.17.0
django-nested-admin>=3.3.3,<3.4.0
psycopg2>=2.7.5,<2.8.0

flake8>=3.6.0,<3.7.0
